requirements.txt
pandas
numpy
scikit-learn
shap
mlflow
matplotlib
scipy
faiss-cpu
sentence-transformers
python-multipart
google-genai
fastapi
uvicorn
pydantic
openpyxl
